Added code to determine binsize for radial profile histogram.
Before committing this, I also tested just fitting the radial profile;
no go - not unexpected, since the fit would be trying to account for
all the non-averaged data. In the end, I'm letting the user define the
binsize, or (else) using astroML to determine the optimum binsize using
the 'knuth' method, or (else) if astroML is not present, a default
value is used.